Refactored Common. Deleted redundant code. Unified all the infrastructure.
