--- src/DShop.Common/Builders/IBusServiceBuilder.cs
@@ -1,10 +1,0 @@
-using System;
-using DShop.Common.Bus;
-
-namespace DShop.Common.Builders
-{
-    public interface IBusServiceBuilder
-    {
-         IServiceBuilder WithServiceBus(string settingsSectionName, Action<ISubscribeBus> subscriptions);
-    }
-}
