--- src/DShop.Common/Builders/IDatabaseServiceBuilder.cs
@@ -1,8 +1,0 @@
-namespace DShop.Common.Builders
-{
-    public interface IDatabaseServiceBuilder
-    {
-        IBusServiceBuilder WithMongoDb(string settingsSectionName);
-        IBusServiceBuilder WithNoDatabase();
-    }
-}
