--- src/DShop.Common/Builders/IIoCServiceBuilder.cs
@@ -1,10 +1,0 @@
-using System;
-using Autofac;
-
-namespace DShop.Common.Builders
-{
-    public interface IIoCServiceBuilder
-    {
-        IDatabaseServiceBuilder WithAutofac(Action<ContainerBuilder> registerDependencies);
-    }
-}
