--- src/DShop.Common/Builders/IServiceBuilder.cs
@@ -1,10 +1,0 @@
-using Microsoft.AspNetCore.Hosting;
-
-namespace DShop.Common.Builders
-{
-    public interface IServiceBuilder
-    {
-        IIoCServiceBuilder WithPort(int port);
-        IWebHost Build();
-    }
-}
