--- src/DShop.Common/Bus/RabbitMq/IBusPublisher.cs
@@ -1,12 +1,0 @@
-using System.Threading.Tasks;
-using DShop.Messages.Commands;
-using DShop.Messages.Events;
-
-namespace DShop.Common.Bus.RabbitMq
-{
-    public interface IBusPublisher
-    {
-        Task PublishCommandAsync<T>(T command) where T : ICommand;
-        Task PublishEventAsync<T>(T @event) where T : IEvent;
-    }
-}
