--- src/DShop.Common/IoC/IDependencyResolver.cs
@@ -1,7 +1,0 @@
-namespace DShop.Common.IoC
-{
-    internal interface IDependencyResolver
-    {
-         TInstance Resolve<TInstance>();
-    }
-}
