+++ src/DShop.Common/Mongo/Extensions.cs
@@ -1,11 +1,9 @@
 using Autofac;
 using DShop.Common.Options;
 using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Options;
 using MongoDB.Driver;
 
-namespace DShop.Common.Databases.Mongo
+namespace DShop.Common.Mongo
 {
     public static class Extensions
     {
@@ -14,34 +12,33 @@ namespace DShop.Common.Databases.Mongo
             builder.Register(context =>
             {
                 var configuration = context.Resolve<IConfiguration>();
-                var section = configuration.GetSection("mongo");
-                var options = new MongoDbOptions();
-                section.Bind(options);
-
+                var options = configuration.GetOptions<MongoDbOptions>("mongo");
                 return options;
+
             }).SingleInstance();
 
-            builder.Register(context => 
-            {
-                var options = context.Resolve<MongoDbOptions>();
-
-                return new MongoClient(options.ConnectionString);
-            }).SingleInstance();
             builder.Register(context =>
             {
                 var options = context.Resolve<MongoDbOptions>();
-                var client = context.Resolve<MongoClient>();     
-                      
+                return new MongoClient(options.ConnectionString);
+
+            }).SingleInstance();
+
+            builder.Register(context =>
+            {
+                var options = context.Resolve<MongoDbOptions>();
+                var client = context.Resolve<MongoClient>();
                 return client.GetDatabase(options.Database);
+
             }).InstancePerLifetimeScope();
 
             builder.RegisterType<MongoDbInitializer>()
                 .As<IMongoDbInitializer>()
                 .InstancePerLifetimeScope();
-                
+
             builder.RegisterType<MongoDbSeeder>()
                 .As<IMongoDbSeeder>()
                 .InstancePerLifetimeScope();
-        }            
+        }
     }
 }
