+++ src/DShop.Common/Mongo/MongoDbOptions.cs
@@ -1,4 +1,4 @@
-namespace DShop.Common.Databases.Mongo
+namespace DShop.Common.Mongo
 {
     public class MongoDbOptions
     {
