+++ src/DShop.Common/RabbitMq/IBusPublisher.cs
@@ -1,0 +1,12 @@
+using System.Threading.Tasks;
+using DShop.Messages.Commands;
+using DShop.Messages.Events;
+
+namespace DShop.Common.RabbitMq
+{
+    public interface IBusPublisher
+    {
+        Task PublishCommandAsync<TCommand>(TCommand command, ICorrelationContext context = null) where TCommand : ICommand;
+        Task PublishEventAsync<TEvent>(TEvent @event) where TEvent : IEvent;
+    }
+}
